+++ aiorpcx/session.py
@@ -35,7 +35,7 @@ from math import ceil
 import time
 
 from aiorpcx.curio import (
-    TaskGroup, TaskTimeout, CancelledError, timeout_after, sleep
+    TaskGroup, TaskTimeout, timeout_after, sleep
 )
 from aiorpcx.framing import (
     NewlineFramer, BitcoinFramer, BadMagicError, BadChecksumError, OversizedPayloadError
@@ -184,6 +184,11 @@ class SessionBase:
         if abs(self.cost - self._cost_last) > 100:
             self.recalc_concurrency()
 
+    def on_disconnect_due_to_excessive_session_cost(self):
+        '''Called just before disconnecting from the session, if it was consuming too
+        much resources.
+        '''
+
     def recalc_concurrency(self):
         '''Call to recalculate sleeps and concurrency for the session.  Called automatically if
         cost has drifted significantly.  Otherwise can be called at regular intervals if
@@ -306,10 +311,9 @@ class MessageSession(SessionBase):
             self.logger.info(f'incoming request timed out after {timeout} secs')
             self._bump_errors()
         except ExcessiveSessionCostError:
+            self.on_disconnect_due_to_excessive_session_cost()
             await self.close()
-        except CancelledError:
-            raise
-        except Exception:
+        except Exception:   # pylint:disable=W0703
             self.logger.exception(f'exception handling {message}')
             self._bump_errors()
 
@@ -474,11 +478,10 @@ class RPCSession(SessionBase):
             result = e.args[0]
             disconnect = True
         except ExcessiveSessionCostError:
+            self.on_disconnect_due_to_excessive_session_cost()
             result = RPCError(JSONRPC.EXCESSIVE_RESOURCE_USAGE, 'excessive resource usage')
             disconnect = True
-        except CancelledError:
-            raise
-        except Exception:
+        except Exception:     # pylint:disable=W0703
             self.logger.exception(f'exception handling {request}')
             result = RPCError(JSONRPC.INTERNAL_ERROR, 'internal server error')